html/browsers/browsing-the-web/back-forward-cache/events.html

<!DOCTYPE HTML>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/resources/channel.sub.js"></script>
<script src="resources/helper.sub.js"></script>
<script>
for (const originType of ['SameOrigin', 'SameSite', 'CrossSite']) {
  promise_test(async t => {
    const remote = new TestDriverRemote();

    window.open(`${executorPath}${remote.uuid}&events=pagehide,pageshow,load`,
      '_blank', 'noopener');

    t.add_cleanup(async () => {
      await remote.executeScript(() => window.close());
      await remote.close();
    });

    // Navigate to a page that immediately triggers back navigation.
    const backUrl = origins[originType] + backPath;

    await remote.postMessage({name: "navigateTo", url: backUrl})
    // Close the reader channel on the remote page so it doesn't read subsequent commands
    remote.pause()

    // The channels will reconnect after navigation
    await assert_bfcached(remote);

    let recordedEvents = await remote.executeScript(() => getRecordedEvents());
    assert_array_equals(
      recordedEvents.toLocal(),
      [
        'window.load',
        'window.pageshow',
        'window.pagehide.persisted',
        'window.pageshow.persisted'
      ]);
  }, `Events fired (window.open + noopener, ${originType})`);
}
</script>
